+++ common/src/main/java/farm/rosehearth/compatemon/mixin/client/MixinNicknameEntryWidget.java
@@ -1,7 +1,9 @@
-package farm.rosehearth.compatemon.mixin;
+package farm.rosehearth.compatemon.mixin.client;
 
 import com.cobblemon.mod.common.client.gui.summary.widgets.NicknameEntryWidget;
 import com.cobblemon.mod.common.pokemon.Pokemon;
+import farm.rosehearth.compatemon.Compatemon;
+import farm.rosehearth.compatemon.modules.compatemon.CompatemonConfig;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.network.chat.Component;
@@ -9,6 +11,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import static farm.rosehearth.compatemon.util.CompatemonDataKeys.MOD_ID_COMPATEMON;
 
 @Mixin(NicknameEntryWidget.class)
 public class MixinNicknameEntryWidget extends EditBox {
@@ -20,6 +24,16 @@ public class MixinNicknameEntryWidget extends EditBox {
 		super(font, x, y, width, height, message);
 	}
 	
+	@ModifyConstant(method="<init>"
+			,constant = @Constant(intValue=12)
+			,remap=false)
+	public int compatemon$fixMaxNameLength(int value){
+		if(Compatemon.ShouldLoadMod(MOD_ID_COMPATEMON))
+			return CompatemonConfig.NicknameFieldLength;
+		return 12;
+	}
+
+
 //	@Inject(at=@At("RETURN")
 //	,remap=false
 //	,method="<init>")
@@ -27,21 +41,12 @@ public class MixinNicknameEntryWidget extends EditBox {
 //
 //	}
 //
-	@ModifyConstant(method="<init>"
-			,constant = @Constant(intValue=12)
-			,remap=false)
-	public int compatemon$fixMaxNameLength(int value){
-		return 32;
-	}
-	
-	
 //	@Inject(at=@At("RETURN")
 //			,remap=false
 //			,method="setSelectedPokemon")
 //	public void compatemon$setSelectedPokemon(Pokemon pokemon, CallbackInfo cir){
 //		super.setMessage(pokemon.getDisplayName());
 //	}
-	
 //	@Inject(at=@At("RETURN")
 //			,remap=false
 //			,method="updateNickname")